(PC-15745) refactor(cheatcodes): Use common LinkToComponent for cheatcode navigation
