+++ src/features/cheatcodes/pages/NavigationSignUp/NavigationSignUp.tsx
@@ -3,17 +3,13 @@ import React, { useState } from 'react'
 import { ScrollView } from 'react-native'
 import styled from 'styled-components/native'
 
+import { LinkToComponent } from 'features/cheatcodes/components/LinkToComponent'
 import { FastEduconnectConnectionRequestModal } from 'features/identityCheck/components/FastEduconnectConnectionRequestModal'
-import {
-  RootScreenNames,
-  RootStackParamList,
-  UseNavigationType,
-} from 'features/navigation/RootNavigator'
+import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { useGoBack } from 'features/navigation/useGoBack'
-import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ModalHeader } from 'ui/components/modals/ModalHeader'
 import { ArrowPrevious } from 'ui/svg/icons/ArrowPrevious'
-import { padding, Spacer } from 'ui/theme'
+import { Spacer } from 'ui/theme'
 
 export function NavigationSignUp(): JSX.Element {
   const { goBack } = useGoBack('Navigation', undefined)
@@ -85,36 +81,8 @@ export function NavigationSignUp(): JSX.Element {
   )
 }
 
-interface LinkToComponentProps {
-  name?: RootScreenNames
-  onPress?: () => void
-  title?: string
-  navigationParams?: RootStackParamList[RootScreenNames]
-}
-
-const LinkToComponent = ({
-  name = 'NavigationSignUp',
-  onPress,
-  title,
-  navigationParams,
-}: LinkToComponentProps) => {
-  const { navigate } = useNavigation<UseNavigationType>()
-  const navigateToComponent = () => navigate(name, navigationParams)
-
-  return (
-    <Row half>
-      <ButtonPrimary wording={title ?? name} onPress={onPress ?? navigateToComponent} />
-    </Row>
-  )
-}
-
 const StyledContainer = styled.View({
   display: 'flex',
   flexWrap: 'wrap',
   flexDirection: 'row',
 })
-
-const Row = styled.View<{ half?: boolean }>(({ half = false }) => ({
-  width: half ? '50%' : '100%',
-  ...padding(2, 0.5),
-}))
